Bug 1810565 - Move logins related UI tests to new class
